requirements.txt
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.7.8
google-cloud-aiplatform==1.38.1
cohere==4.39
requests==2.31.0
aiohttp==3.9.3
asyncio==3.4.3
fastapi==0.109.2
uvicorn==0.27.1
pydantic==1.10.13
python-multipart==0.0.6
